# apps/web/src/generated/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Secondlayer Public API",
		"version": "1.0.0",
		"description": "This instance: Index (decoded chain events), Streams (raw firehose), Subgraphs (your schema). Cursor format is `<block_height>:<event_index>` on Index/Streams; opaque on Subgraphs."
	},
	"servers": [
		{
			"url": "/",
			"description": "This instance"
		}
	],
...
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "sk-sl_*"
			}
		},
		"schemas": {
...
			"Error": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string"
					},
					"code": {
						"type": "string"
					}
				}
			}
		},
...
	},
	"paths": {
...
		"/v1/openapi.json": {
			"get": {
				"summary": "This document",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/v1/subgraphs": {
			"get": {
				"tags": [
					"subgraphs"
				],
				"summary": "List readable subgraphs (public + your own with a bearer key)",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
...
		"/v1/index": {
			"get": {
				"tags": [
					"index"
				],
				"summary": "Index discovery",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
...
		"/v1/streams": {
			"get": {
				"tags": [
					"streams"
				],
				"summary": "Streams discovery",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
...
	}
}